backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Session==0.8.0
flatbuffers==24.3.25
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.67.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
msgspec==0.18.6
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.52.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.13.1
setuptools==75.2.0
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.16.0
XlsxWriter==3.2.0
